refactor(protocol): moved protocol from transport to protocol crate
