+++ src/protocol/src/lib.rs
@@ -1,0 +1,5 @@
+//! # Knot Protocol
+//!
+//! This crate provides the communication protocol for the Knot daemon.
+
+pub mod daemon;
